unify components of common module
