+++ lookbook/previews/open_project/common/loading_indicator_preview.rb
@@ -1,0 +1,8 @@
+module OpenProject
+  module Common
+    # @logical_path OpenProject/Common
+    class LoadingIndicatorPreview < Lookbook::Preview
+      def default; end
+    end
+  end
+end
