+++ lookbook/previews/open_project/common/loading_indicator_preview/default.html.erb
