--- lookbook/previews/open_project/loading_indicator_preview.rb
@@ -1,6 +1,0 @@
-module OpenProject
-  # @logical_path OpenProject/common
-  class LoadingIndicatorPreview < Lookbook::Preview
-    def default; end
-  end
-end
